requirements.txt
requests==2.31.0
pillow==10.2.0
beautifulsoup4==4.12.3
selenium==4.18.1
anthropic==0.18.1
python-dotenv==1.0.1